+++ packages/design-system-importer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @auto-engineer/design-system-importer
 
+## 0.9.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @auto-engineer/message-bus@0.9.7
+
 ## 0.9.6
 
 ### Patch Changes