+++ packages/file-store/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @auto-engineer/file-store
 
+## 0.9.7
+
 ## 0.9.6
 
 ## 0.9.5